tabulate

tji asd asd sda sad sa d tji asd asd sda sad sa tji asd asd sda sad sa tji asd asd sda sad sa tji asd asd sda sad sa tji asd asd sda sad sa tji asd asd sda sad sa tji asd asd sda sad sa #'

tabulate(
  data,
  cols,
  weights = NULL,
  groups = NULL,
  samples = NULL,
  return_mean = FALSE,
  values_drop_na = TRUE,
  variable_sep = NULL,
  variable_sep_suffix = NULL,
  keep_empty_levels = FALSE
)

Arguments

data

A dataframe

cols

[str] string vector of variables or tidy supported selection ex. matches(...)

weights

[str] string vector of variables or tidy supported selection ex. matches(...)

groups

[str] string vector of variables or tidy supported selection ex. matches(...)

samples

[str] string vector of variables or tidy supported selection ex. matches(...)

return_mean

[boolean] for nominal variables

values_drop_na

[boolean] if the base and pct should include NAs

variable_sep

[str] regex used to separate the variable column

variable_sep_suffix

[str] name of the variable_sep right-hand side output column

keep_empty_levels

[str] no levels imputations

Value

frequency table, dataframe/tibble, A standardized ET

Examples

tabulate(iris, Species)
#> # A tibble: 3 × 6
#>   weight     variable value          n  base   pct
#>   <chr>      <chr>    <chr>      <dbl> <dbl> <dbl>
#> 1 unweighted Species  setosa        50   150 0.333
#> 2 unweighted Species  versicolor    50   150 0.333
#> 3 unweighted Species  virginica     50   150 0.333